requirements.txt
openpyxl
streamlit
odfpy
pandas
pandas
matplotlib